Refactor compiler
